+++ lib/ex_check/check.ex
@@ -263,6 +263,10 @@ defmodule ExCheck.Check do
     ["missing directory ", b(cd)]
   end
 
+  defp format_skip_reason(:no_fix) do
+    ["no fix"]
+  end
+
   defp format_tool_name(name) when is_atom(name) do
     b(name)
   end